# openapi.yaml
openapi: 3.0.3
info:
  title: Karton API
  description: >
    API za aplikaciju "Karton" — vođenje auto servisa. Backend je jedini autoritet
    za poslovna pravila (specifikacija-finalna.md §7). Nazivi na engleskom (odluka 27),
    poruke grešaka na srpskom. Sva vremena u UTC; poslovna logika Europe/Belgrade.

    CSRF ugovor: svaki odgovor postavlja/osvežava XSRF-TOKEN cookie (čitljiv iz JS);
    svaki non-GET zahtev MORA poslati header X-CSRF-Token sa tom vrednošću —
    neslaganje ili izostanak → 403 CSRF_FAILED. Važi i za /auth/login (token se
    dobija prethodnim GET pozivom, npr. /auth/me).

    Optimističko zaključavanje: verzionisani agregati su work_order, document,
    appointment i settings. Sve mutacione akcije nad njima nose "version" u telu;
    zastarela verzija → 409 VERSION_CONFLICT. Akcije bez version polja (copy,
    DELETE termina, izdavanje predračuna) backend izvršava u transakciji uz
    row-lock i atomsku proveru aktuelnog stanja.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /customers:
    get:
      tags: [customers]
      summary: Lista klijenata (server-side filter/sort/paginacija)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: type, in: query, schema: { type: string, enum: [individual, company] } }
        - { name: status, in: query, schema: { type: string, enum: [active, archived] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerPage' } } } }
    post:
      tags: [customers]
      summary: Novi klijent (PIB obavezan za pravna lica; duplikat 409)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerInput' } } } }
      responses:
        '201': { description: Kreiran, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '409': { $ref: '#/components/responses/Duplicate' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /customers/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [customers]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } } }
    patch:
      tags: [customers]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CustomerInput' } } } }
      responses:
        '200': { description: Izmenjen, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /customers/{id}/contacts:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [customers]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Contact' } } } } } }
    post:
      tags: [customers]
      summary: Dodaj telefon/email (zamena traži potvrdu na frontu)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ContactInput' } } } }
      responses: { '201': { description: Kreiran } }
  /customers/{id}/archive:
    post:
      tags: [customers]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      summary: Arhiviraj (audit customer.archived)
      responses: { '204': { description: Arhiviran } }
  /customers/{id}/unarchive:
    post:
      tags: [customers]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses: { '204': { description: Dearhiviran } }

  # ---------- Vehicles ----------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: karton_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
...
    Page: { name: page, in: query, schema: { type: integer, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, default: 20 }, description: Default iz Podešavanja }
    Sort: { name: sort, in: query, schema: { type: string }, description: 'npr. name:asc' }
    Query: { name: q, in: query, schema: { type: string }, description: Case-insensitive, ignoriše razmake }
  responses:
...
    Conflict:
      description: VERSION_CONFLICT — optimističko zaključavanje
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Duplicate:
      description: DUPLICATE_VIN / DUPLICATE_TAX_ID / DUPLICATE_EMAIL — nosi existingId
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BusinessRule:
      description: 'Poslovno pravilo (422) — kod iz kataloga grešaka (spec §8)'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: PROFORMA_NOT_VALID }
        message: { type: string, example: 'Predračun je istekao i ne može se pretvoriti u račun.' }
        fields: { type: object, additionalProperties: { type: string } }
        warnings: { type: array, items: { type: string }, example: [MECHANIC_BUSY] }
        existingId: { type: integer, description: Kod duplikata — ID postojećeg entiteta }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [individual, company] }
        name: { type: string }
        taxId: { type: string, nullable: true }
        address: { type: string }
        status: { type: string, enum: [active, archived] }
        contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } }
    CustomerInput:
      type: object
      required: [type, name]
      properties:
        type: { type: string, enum: [individual, company] }
        name: { type: string }
        taxId: { type: string, description: PIB obavezan za company (BR-04) }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
    CustomerPage:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Customer' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
    Contact:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [phone, email] }
        value: { type: string }
        isPrimary: { type: boolean }
    ContactInput:
      type: object
      required: [kind, value]
      properties:
        kind: { type: string, enum: [phone, email] }
        value: { type: string }
        isPrimary: { type: boolean }
...